Reimplement integration test for symlink Dockerfile as a unit test
